pyproject.toml
[tool.poetry]
name = "python-ai-project"
version = "0.1.0"
description = "A reusable AI project starter template"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.0.0"
langchain = "^0.1.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
black = "^24.0"
ruff = "^0.4.0"
pytest = "^8.0"
pyright = "^1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"